Updating the prompting and session setup
Session setup now has a temperature set, as well as better initial prompting and also a check for tokens remaining.

Also adding a quick regex lookup to ignore certain elements on reddit, Ill likely need to expand this.

Adding a character limit to the prompt

Solving a bug in caching as well as a bug in initial prompts